# docs/generated/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "1BIS Backend API",
    "version": "1.0.0",
    "description": "Welcome to the 1BIS API documentation. This is a RESTful API that allows you to interact with the 1BIS backend services.",
    "contact": {
      "name": "1BIS Support",
      "email": "support@1bis.com",
      "url": "https://1bis.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://1bis.com/terms"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    },
    {
      "url": "https://api.1bis.com/api",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/transaction": {
      "get": {
        "summary": "Get all transactions",
        "description": "Retrieve transactions with advanced filtering, pagination, and sorting",
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON array of filter objects"
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create new transaction",
        "description": "Create a new payment transaction",
        "tags": [
          "Transaction"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionNumber",
                  "employeeId",
                  "orderId",
                  "type",
                  "amount",
                  "paymentMethod"
                ],
                "properties": {
                  "transactionNumber": {
                    "type": "string"
                  },
                  "employeeId": {
                    "type": "string"
                  },
                  "orderId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "PURCHASE",
                      "INSTALLMENT",
                      "POINTS_REDEMPTION",
                      "REFUND",
                      "ADJUSTMENT"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "INSTALLMENT",
                      "POINTS",
                      "MIXED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}